refactor: Service 계층 인터페이스 분리
