+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -1,30 +1,13 @@
 package roomescape.time.service;
 
-import org.springframework.stereotype.Service;
 import roomescape.time.domain.ReservationTime;
-import roomescape.time.repository.ReservationTimeRepository;
 
 import java.time.LocalTime;
 import java.util.List;
 
-@Service
-public class ReservationTimeService {
+public interface ReservationTimeService {
 
-    private final ReservationTimeRepository reservationTimeRepository;
-
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
-        this.reservationTimeRepository = reservationTimeRepository;
-    }
-
-    public ReservationTime createTime(LocalTime startAt) {
-        return reservationTimeRepository.save(startAt);
-    }
-
-    public List<ReservationTime> getTimes() {
-        return reservationTimeRepository.findAll();
-    }
-
-    public void removeTime(Long id) {
-        reservationTimeRepository.remove(id);
-    }
+    List<ReservationTime> getTimes();
+    ReservationTime createTime(LocalTime localTime);
+    void removeTime(Long id);
 }
